+++ src/metawards/utils/_add_wards_network_distance.pyx
@@ -6,7 +6,9 @@ cimport cython
 from cython.parallel import parallel, prange
 cimport openmp
 
-from libc.math cimport sqrt
+from libc.math cimport sqrt, sin, cos, atan2, M_PI
+
+import math
 
 from .._parameters import Parameters
 from .._network import Network
@@ -14,6 +16,48 @@ from .._network import Network
 from ._get_array_ptr cimport get_int_array_ptr, get_double_array_ptr
 
 __all__ = ["add_wards_network_distance"]
+
+
+cdef inline double deg_to_rad(double x) nogil:
+    """Convert the passed angle in degrees to radians"""
+    return x * M_PI / 180
+
+
+cdef double distance_lat_long(double lon1, double lat1,
+                              double lon2, double lat2) nogil:
+    """Calculate the distance in kilometers between two lat/lon points on the
+       Earth's surface
+    """
+    cdef double radius = 6378.16  # Earth's radius in km
+    cdef double dlon = deg_to_rad(lon2 - lon1)
+    cdef double dlat = deg_to_rad(lat2 - lat1)
+
+    # Calculate using the haversin formula
+    cdef double sin_dlat_over_2 = sin(dlat / 2.0)
+    cdef double sin_dlon_over_2 = sin(dlon / 2.0)
+
+    cdef double a = (sin_dlat_over_2 * sin_dlat_over_2) + \
+                    (cos(deg_to_rad(lat1)) * cos(deg_to_rad(lat2)) *
+                     sin_dlon_over_2 * sin_dlon_over_2)
+
+    cdef double angle = 2.0 * atan2(sqrt(a), sqrt(1 - a))
+
+    return angle * radius
+
+
+cdef double distance_x_y(double x1, double y1,
+                         double x2, double y2) nogil:
+    """Calculate the distance between two x/y points on
+       a plane, in the units of the points (hopefully they
+       are in kilometers too...)
+    """
+    cdef double dx = x1 - x2
+    cdef double dy = y1 - y2
+
+    dx *= dx
+    dy *= dy
+
+    return sqrt(dx + dy)
 
 
 def add_wards_network_distance(network: Network, nthreads: int = 1):
@@ -41,6 +85,18 @@ def add_wards_network_distance(network: Network, nthreads: int = 1):
     cdef double * wards_x = get_double_array_ptr(wards.x)
     cdef double * wards_y = get_double_array_ptr(wards.y)
 
+    cdef scale_for_km = 1.0
+
+    if params.input_files.coordinates == "x/y":
+        calc_distance = distance_x_y
+        scale_for_km = 0.001
+    elif params.input_files.coordinates == "lat/long":
+        calc_distance = distance_lat_long
+        scale_for_km = 1.0
+    else:
+        raise ValueError(f"Unrecognised coordinate system "
+                         f"{params.input_files.coordinates}")
+
     try:
         with open(params.input_files.position, "r") as FILE:
             line = FILE.readline()
@@ -51,14 +107,39 @@ def add_wards_network_distance(network: Network, nthreads: int = 1):
                 x = float(words[1])
                 y = float(words[2])
 
-                wards.x[i1] = x
-                wards.y[i1] = y
+                wards.x[i1] = x * scale_for_km
+                wards.y[i1] = y * scale_for_km
+
+                if (x == 0) and (y == 0):
+                    print(f"WARNING: Position of ward {i1} is 0,0. This "
+                          f"doesn't seem right")
 
                 line = FILE.readline()
     except Exception as e:
         raise ValueError(f"{params.input_files.position} is corrupted or "
                          f"unreadable? Error = {e.__class__}: {e}, "
                          f"line = {line}")
+
+    # make sure that all nodes have valid distances
+    cdef int i = 0
+    cdef int nmissing = 0
+
+    for i in range(1, network.nnodes+1):
+        x = wards.x[i]
+        y = wards.y[i]
+
+        if (x == 0) and (y == 0):
+            nmissing += 1
+            if nmissing == 20:
+                print("Not printing any more missing positions as there "
+                      "are too many. You should fix the positions file")
+            elif nmissing < 20:
+                print(f"WARNING: Position of ward {i} does not appear to have "
+                      f"been set - position is ({x} {y}).")
+
+    if nmissing > 0:
+        print(f"In total the number of wards with missing positions "
+              f"is {nmissing}")
 
     print("Calculating distances...")
 
@@ -78,16 +159,15 @@ def add_wards_network_distance(network: Network, nthreads: int = 1):
     cdef double x2 = 0.0
     cdef double y1 = 0.0
     cdef double y2 = 0.0
-    cdef double dx = 0.0
-    cdef double dy = 0.0
 
-    cdef int nnodes_plus_one = network.nnodes + 1
     cdef int nlinks_plus_one = network.nlinks + 1
+    cdef int plinks_plus_one = network.plinks + 1
+
+    cdef double too_large = 1000000 * scale_for_km
 
     cdef int ifrom = 0
     cdef int ito = 0
 
-    cdef int i = 0
     cdef int num_threads = nthreads
 
     with nogil, parallel(num_threads=num_threads):
@@ -101,20 +181,28 @@ def add_wards_network_distance(network: Network, nthreads: int = 1):
             x2 = wards_x[ito]
             y2 = wards_y[ito]
 
-            dx = x1 - x2
-            dy = y1 - y2
+            if (x1 == 0 and y1 == 0) or (x2 == 0 and y2 == 0):
+                links_distance[i] = 0.0
+            else:
+                # skipping null points
+                distance = calc_distance(x1, y1, x2, y2)
 
-            distance = sqrt(dx*dx + dy*dy)
-            links_distance[i] = distance
-            total_distance += distance
+                links_distance[i] = distance
+                total_distance += distance
+
+                if distance > too_large:
+                    with gil:
+                        print(f"Large distance between wards {ifrom} "
+                              f"and {ito}: {distance} km? {x1},{y1}  "
+                              f"{x2},{y2}")
 
     print(f"Total links distance equals {total_distance}")
 
-    total_distance = 0.0
+    cdef double total_play_distance = 0.0
 
     if network.plinks > 0:
-        with nogil, parallel(num_threads=num_threads):
-            for i in prange(1, nnodes_plus_one, schedule="static"):
+        with nogil, parallel(num_threads=1):
+            for i in prange(1, plinks_plus_one, schedule="static"):
                 ifrom = plinks_ifrom[i]
                 ito = plinks_ito[i]
 
@@ -124,13 +212,22 @@ def add_wards_network_distance(network: Network, nthreads: int = 1):
                 x2 = wards_x[ito]
                 y2 = wards_y[ito]
 
-                dx = x1 - x2
-                dy = y1 - y2
+                if (x1 == 0 and y1 == 0) or (x2 == 0 and y2 == 0):
+                    plinks_distance[i] = 0.0
+                else:
+                    # skipping null points
+                    distance = calc_distance(x1, y1, x2, y2)
 
-                distance = sqrt(dx*dx + dy*dy)
-                plinks_distance[i] = distance
-                total_distance += distance
+                    plinks_distance[i] = distance
+                    total_play_distance += distance
 
-        print(f"Total play links distance equals {total_distance}")
+                    if distance > too_large:
+                        with gil:
+                            print(f"Large distance between play wards {ifrom} "
+                                  f"and {ito}: {distance} km? {x1},{y1}  "
+                                  f"{x2},{y2}")
+
+    print(f"Total play distance equals {total_play_distance}")
+    print(f"Total distance equals {total_distance+total_play_distance}")
 
     return network