+++ src/metawards/utils/_build_play_matrix.pyx
@@ -108,7 +108,7 @@ def build_play_matrix(network: Network,
                 links_ito[nlinks] = to_id
                 links_weight[nlinks] = weight
 
-                nodes_denominator_p[from_id] += weight  # not denominator_p
+                nodes_denominator_p[from_id] += weight
                 nodes_play_suscept[from_id] += weight
 
             fclose(cfile)
@@ -129,6 +129,9 @@ def build_play_matrix(network: Network,
     p = p.start("renormalise?")
     renormalise = (params.input_files.play == params.input_files.work)
     p = p.stop()
+
+    if renormalise:
+        print("Identical work and play links, so renormalising...")
 
     p = p.start("renormalise_loop")
     for j in range(1, nlinks+1):   # careful 1-indexed